refactor: move WhatsApp models and client to smsclient/termii
Consolidate WhatsApp-related code (models and client) under the smsclient/termii package for better organization and logical grouping. This removes the previous split where models were in emailclient/termii and the client was missing, and updates all import references accordingly.
