+++ lib/smsclient/termii/models.go
@@ -70,3 +70,21 @@ func (v *verifyTokenResponse) UnmarshalJSON(data []byte) error {
 
 	return nil
 }
+
+type whatsAppRequest struct {
+	APIKey  string `json:"api_key"`
+	To      string `json:"to"`
+	From    string `json:"from"`
+	SMS     string `json:"sms"`
+	Channel string `json:"channel"`
+	Type    string `json:"type"`
+}
+
+type whatsAppResponse struct {
+	Code         string  `json:"code"`
+	Balance      float64 `json:"balance"`
+	MessageID    string  `json:"message_id"`
+	Message      string  `json:"message"`
+	User         string  `json:"user"`
+	MessageIDStr string  `json:"message_id_str"`
+}
